Add image based on Fedora 36
Add a Dockerfile for an image based on Fedora 36, containing gcc 12,
qemu, and other tools needed to run the EDK2 CI.

Also put GitHub Action in place that builds the image and pushes it to
the GitHub image registry (ghcr.io). The image shows up under "Packages"
on the GitHub page as well. The image is only pushed for changes on the
main branch. Pull Requests build the image, but do not push it.